Add files via upload
The aim of this project is to build a Quarantine center management system which works on the entries given by all the visited persons and retrieve the details based on query. This project uses almost all the important concepts of c++ and Object Oriented Programming to manage the Quarantine centre in a best possible way.